+++ lib/libs/shared/src/types/loyalty-service/iloyalty-reward.d.ts
@@ -4,19 +4,19 @@ import { BaseResponse } from '../common.type';
 import { LOYALTY_REWARD_DISCOUNT_APPLICABILITY_ENUM, LOYALTY_REWARD_DISCOUNT_MODE_ENUM, LOYALTY_REWARD_TYPE_ENUM } from '../../enum/loyalty-service';
 export declare namespace ILoyaltyReward {
     interface ILoyaltyRewardCreate {
-        program_id?: number & tags.Type<'uint32'>;
+        program_id?: number & tags.Type<'uint64'>;
         active?: boolean;
         description?: string;
         rewardType: string & LOYALTY_REWARD_TYPE_ENUM;
         discount?: number;
         discountMode?: string & LOYALTY_REWARD_DISCOUNT_MODE_ENUM;
         discountApplicability?: string & LOYALTY_REWARD_DISCOUNT_APPLICABILITY_ENUM;
-        discount_product_category_id?: number & tags.Type<'uint32'>;
-        discount_product_tag_id?: number & tags.Type<'uint32'>;
+        discount_product_category_id?: number & tags.Type<'uint64'>;
+        discount_product_tag_id?: number & tags.Type<'uint64'>;
         discountMaxAmount?: number;
-        discount_line_product_id?: number & tags.Type<'uint32'>;
-        reward_product_id?: number & tags.Type<'uint32'>;
-        reward_product_tag_id?: number & tags.Type<'uint32'>;
+        discount_line_product_id?: number & tags.Type<'uint64'>;
+        reward_product_id?: number & tags.Type<'uint64'>;
+        reward_product_tag_id?: number & tags.Type<'uint64'>;
         rewardProductQty?: number;
         requiredPoints?: number;
         rewardProducts?: ILoyaltyRewardProduct.ILoyaltyRewardProductCreate[] | null;
