+++ lib/libs/shared/src/types/payment-service/ipayment-gateway.type.d.ts
@@ -2,7 +2,7 @@ import { tags } from 'typia';
 import { BaseResponse } from '../common.type';
 export declare namespace IPaymentGateway {
     interface ICreatePaymentGateway {
-        shop_id?: number & tags.Type<'uint32'>;
+        shop_id?: number & tags.Type<'uint64'>;
         name: string;
         abbreviation: string;
         code: string;
