+++ lib/libs/shared/src/types/product-service/itag.d.ts
@@ -4,35 +4,35 @@ export declare namespace ITag {
     interface IReqCreateTag {
         name: string;
         icon?: string | null | undefined;
-        parent_id?: number & tags.Type<'uint32'>;
-        image_id?: number & tags.Type<'uint32'>;
+        parent_id?: number & tags.Type<'uint64'>;
+        image_id?: number & tags.Type<'uint64'>;
         userId?: string;
     }
     interface ICreateTag extends IReqCreateTag {
-        shop_id?: number & tags.Type<'uint32'>;
+        shop_id?: number & tags.Type<'uint64'>;
     }
     interface IOnCreateTag {
         name: string;
         icon?: string | null | undefined;
-        shopId: number & tags.Type<'uint32'>;
-        imageId?: number & tags.Type<'uint32'>;
-        parentId?: number & tags.Type<'uint32'>;
+        shopId: number & tags.Type<'uint64'>;
+        imageId?: number & tags.Type<'uint64'>;
+        parentId?: number & tags.Type<'uint64'>;
         userId?: string;
     }
     interface IUpdateTag {
         name?: string;
         active?: boolean;
         icon?: string | null | undefined;
-        parent_id?: number & tags.Type<'uint32'>;
-        image_id?: number & tags.Type<'uint32'>;
+        parent_id?: number & tags.Type<'uint64'>;
+        image_id?: number & tags.Type<'uint64'>;
         userId?: string;
     }
     interface IOnUpdateTag {
         name?: string;
         active?: boolean;
         icon?: string | null | undefined;
-        imageId?: number & tags.Type<'uint32'>;
-        parentId?: number & tags.Type<'uint32'>;
+        imageId?: number & tags.Type<'uint64'>;
+        parentId?: number & tags.Type<'uint64'>;
         userId?: string;
     }
     interface IDetailTagResponse extends BaseResponse {
